requirements.txt
numpy==1.19.5
paho_mqtt==1.6.1
pandas==1.1.5
requests==2.18.4
redis==3.5.3

